+++ src/Controller/PagesController.php
@@ -69,7 +69,7 @@ class PagesController extends AppController
         $slug = rtrim($slug, '/');
         $page = $this->Pages->findActiveBySlug($slug)
             ->contain([$this->Pages->Categories->getAlias() => ['fields' => ['title', 'slug']]])
-            ->cache(sprintf('view_%s', md5($slug)), $this->Pages->getCacheName())
+            ->cache('view_' . md5($slug))
             ->firstOrFail();
 
         $this->set(compact('page'));