+++ src/Model/Table/UsersTable.php
@@ -14,12 +14,12 @@ namespace MeCms\Model\Table;
 
 use ArrayObject;
 use Cake\Event\Event;
-use Cake\ORM\Query;
 use Cake\ORM\ResultSet;
 use Cake\ORM\RulesChecker;
 use MeCms\Model\Entity\User;
 use MeCms\Model\Table\AppTable;
 use MeCms\Model\Validation\UserValidator;
+use MeCms\ORM\Query;
 
 /**
  * Users model
@@ -131,7 +131,7 @@ class UsersTable extends AppTable
                     return $user->get('first_name') . ' ' . $user->get('last_name');
                 });
             })
-            ->cache(sprintf('active_%s_list', $this->getTable()), $this->getCacheName());
+            ->cache(sprintf('active_%s_list', $this->getTable()));
     }
 
     /**