+++ tests/Fixture/Core/ArticlesFixture.php
@@ -1,0 +1,42 @@
+<?php
+/**
+ * This file is part of me-cms.
+ *
+ * Licensed under The MIT License
+ * For full copyright and license information, please see the LICENSE.txt
+ * Redistributions of files must retain the above copyright notice.
+ *
+ * @copyright   Copyright (c) Mirko Pagliai
+ * @link        https://github.com/mirko-pagliai/me-cms
+ * @license     https://opensource.org/licenses/mit-license.php MIT License
+ */
+namespace MeCms\Test\Fixture\Core;
+
+use Cake\TestSuite\Fixture\TestFixture;
+
+/**
+ * Articles fixture
+ */
+class ArticlesFixture extends TestFixture
+{
+    /**
+     * @var array
+     */
+    public $fields = [
+        'id' => ['type' => 'integer'],
+        'author_id' => ['type' => 'integer', 'null' => true],
+        'title' => ['type' => 'string', 'null' => true],
+        'body' => 'text',
+        'published' => ['type' => 'string', 'length' => 1, 'default' => 'N'],
+        '_constraints' => ['primary' => ['type' => 'primary', 'columns' => ['id']]]
+    ];
+
+    /**
+     * @var array
+     */
+    public $records = [
+        ['author_id' => 1, 'title' => 'First Article', 'body' => 'First Article Body', 'published' => 'Y'],
+        ['author_id' => 3, 'title' => 'Second Article', 'body' => 'Second Article Body', 'published' => 'Y'],
+        ['author_id' => 1, 'title' => 'Third Article', 'body' => 'Third Article Body', 'published' => 'Y']
+    ];
+}